requirements.txt
gevent
Jinja2==3.1.2
psycopg2-binary
orator
